+++ MLEM/Graphics/ColorExtensions.cs
@@ -58,6 +58,39 @@ namespace MLEM.Graphics {
             return $"{(hash ? "#" : string.Empty)}{color.R:X2}{color.G:X2}{color.B:X2}";
         }
 
+        /// <summary>
+        /// Converts the given <paramref name="color"/> into HSL representation and returns the result as a value tuple with values between 0 and 1.
+        /// </summary>
+        /// <remarks>This code is adapted from https://gist.github.com/mjackson/5311256.</remarks>
+        /// <param name="color">The color to convert.</param>
+        /// <returns>The resulting HSL color as a value tuple containing H, S and L, each between 0 and 1.</returns>
+        public static (float H, float S, float L) ToHsl(this Color color) {
+            var r = color.R / 255F;
+            var g = color.G / 255F;
+            var b = color.B / 255F;
+
+            var max = Math.Max(Math.Max(r, g), b);
+            var min = Math.Min(Math.Min(r, g), b);
+            float h, s, l = (max + min) / 2;
+
+            if (max == min) {
+                h = s = 0; // achromatic
+            } else {
+                var d = max - min;
+                s = l > 0.5F ? d / (2 - max - min) : d / (max + min);
+                if (r == max) {
+                    h = (g - b) / d + (g < b ? 6 : 0);
+                } else if (g == max) {
+                    h = (b - r) / d + 2;
+                } else {
+                    h = (r - g) / d + 4;
+                }
+                h /= 6;
+            }
+
+            return (h, s, l);
+        }
+
     }
 
     /// <summary>
@@ -114,6 +147,55 @@ namespace MLEM.Graphics {
             }
             color = default;
             return false;
+        }
+
+        /// <summary>
+        ///Converts the given HSL color to an RGB <see cref="Color"/> and returns the result.
+        /// </summary>
+        /// <remarks>This code is adapted from https://gist.github.com/mjackson/5311256.</remarks>
+        /// <param name="color">The HSL color to convert, as a value tuple that contains H, S and L values, each between 0 and 1.</param>
+        /// <returns>The resulting RGB color.</returns>
+        public static Color FromHsl((float H, float S, float L) color) {
+            return ColorHelper.FromHsl(color.H, color.S, color.L);
+        }
+
+        /// <summary>
+        ///Converts the given HSL values to an RGB <see cref="Color"/> and returns the result.
+        /// </summary>
+        /// <remarks>This code is adapted from https://gist.github.com/mjackson/5311256.</remarks>
+        /// <param name="h">The H component of the HSL color, between 0 and 1.</param>
+        /// <param name="s">The S component of the HSL color, between 0 and 1.</param>
+        /// <param name="l">The L component of the HSL color, between 0 and 1.</param>
+        /// <returns>The resulting RGB color.</returns>
+        public static Color FromHsl(float h, float s, float l) {
+            float r, g, b;
+
+            if (s == 0) {
+                r = g = b = l; // achromatic
+            } else {
+                var q = l < 0.5F ? l * (1 + s) : l + s - l * s;
+                var p = 2 * l - q;
+
+                r = ColorHelper.HueToRgb(p, q, h + 1 / 3F);
+                g = ColorHelper.HueToRgb(p, q, h);
+                b = ColorHelper.HueToRgb(p, q, h - 1 / 3F);
+            }
+
+            return new Color(r, g, b);
+        }
+
+        private static float HueToRgb(float p, float q, float t) {
+            if (t < 0)
+                t += 1;
+            if (t > 1)
+                t -= 1;
+            if (t < 1 / 6F)
+                return p + (q - p) * 6 * t;
+            if (t < 1 / 2F)
+                return q;
+            if (t < 2 / 3F)
+                return p + (q - p) * (2 / 3F - t) * 6;
+            return p;
         }
 
     }